Feature/authorisation

+++ controllers/card_controller.py
@@ -6,6 +6,8 @@ from flask_jwt_extended import jwt_required, get_jwt_identity
 from init import db
 from models.card import Card, card_schema, cards_schema
 from controllers.comment_controller import comments_bp
+from models.user import User
+from utils import authorise_as_admin
 
 
 cards_bp = Blueprint("cards", __name__, url_prefix="/cards")
@@ -61,19 +63,26 @@ def create_card():
     # respond
     return card_schema.dump(card)
 
+# /cards/<id> - DELETE - delete a card
 @cards_bp.route("/<int:card_id>", methods=["DELETE"])
 @jwt_required()
 def delete_card(card_id):
-    # fetch card from database
-    stmt = db.select(Card).filter_by(id = card_id)
+    # fetch the card from the database
+    stmt = db.select(Card).filter_by(id=card_id)
     card = db.session.scalar(stmt)
-    
-    # if card exists
+
+    # if card
     if card:
+        # check whether the user is an admin or not
+        is_admin = authorise_as_admin()
+
+        if not is_admin and str(card.user_id) != get_jwt_identity():
+            return {"error": "User is not authorised to perform this action."}, 403
+        
         # delete the card
         db.session.delete(card)
         db.session.commit()
-
+        
         return {"message": f"Card '{card.title}' deleted successfully"}
     # else
     else:
@@ -91,6 +100,10 @@ def update_card(card_id):
 
     # if card exist
     if card:
+        # if the user is the owner
+        if str(card.user_id) != get_jwt_identity():
+            return {"error": "You are not the owner of the card"}, 403
+        
         # update the fields
         card.title = body_data.get("title") or card.title,
         card.description = body_data.get("description") or card.description,
